Split off also tests.
